# be/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CompanyResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      logo:
        type: string
      name:
        type: string
      npwp:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateCompanyRequest:
    properties:
      address:
        type: string
      email:
        type: string
      logo:
        type: string
      name:
        type: string
      npwp:
        type: string
      phone:
        type: string
    required:
    - name
    type: object
...
  dto.CreateHolidayRequest:
    properties:
      company_id:
        type: string
      date:
        type: string
      is_national:
        type: boolean
      name:
        type: string
    required:
    - company_id
    - date
    - name
    type: object
...
  dto.HolidayResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      is_national:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateCompanyRequest:
    properties:
      address:
        type: string
      email:
        type: string
      is_active:
        type: boolean
      logo:
        type: string
      name:
        type: string
      npwp:
        type: string
      phone:
        type: string
    type: object
...
  dto.UpdateHolidayRequest:
    properties:
      date:
        type: string
      is_national:
        type: boolean
      name:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hris.com
    name: API Support
  description: Human Resource Information System API with authentication and comprehensive
    HRIS modules
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HRIS API
  version: "1.0"
paths:
...
  /companies:
    get:
      description: Retrieve all companies
      produces:
      - application/json
      responses:
        "200":
          description: Companies retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.CompanyResponse'
                  type: array
              type: object
        "500":
          description: Failed to fetch companies
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get all companies
      tags:
      - Companies
    post:
      consumes:
      - application/json
      description: Create a new company
      parameters:
      - description: Company data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCompanyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Company created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a company
      tags:
      - Companies
  /companies/{id}:
    delete:
      description: Delete a company by ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Company deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to delete
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a company
      tags:
      - Companies
    get:
      description: Retrieve a company by its ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Company retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "404":
          description: Company not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get company by ID
      tags:
      - Companies
    put:
      consumes:
      - application/json
      description: Update an existing company by ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      - description: Company data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Company updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CompanyResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a company
      tags:
      - Companies
...
  /holidays:
    get:
      description: Retrieve all holidays, optionally filtered by company and/or year
      parameters:
      - description: Filter by company ID
        in: query
        name: company_id
        type: string
      - description: Filter by year
        in: query
        name: year
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Holidays retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.HolidayResponse'
                  type: array
              type: object
        "400":
          description: Invalid year format
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to fetch holidays
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get all holidays
      tags:
      - Holidays
    post:
      consumes:
      - application/json
      description: Create a new holiday record
      parameters:
      - description: Holiday data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateHolidayRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Holiday created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HolidayResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a holiday
      tags:
      - Holidays
  /holidays/{id}:
    delete:
      description: Delete a holiday by ID
      parameters:
      - description: Holiday ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Holiday deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to delete
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a holiday
      tags:
      - Holidays
    get:
      description: Retrieve a holiday by its ID
      parameters:
      - description: Holiday ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Holiday retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HolidayResponse'
              type: object
        "404":
          description: Holiday not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get holiday by ID
      tags:
      - Holidays
    put:
      consumes:
      - application/json
      description: Update an existing holiday by ID
      parameters:
      - description: Holiday ID
        in: path
        name: id
        required: true
        type: string
      - description: Holiday data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateHolidayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Holiday updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.HolidayResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a holiday
      tags:
      - Holidays
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"